tests/wpt/tests/svg/painting/parsing/stroke-linecap-valid.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="800px">
  <title>SVG Painting: parsing stroke-linecap with valid values</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/painting.html#StrokeLinecapProperty"/>
    <h:meta name="assert" content="stroke-linecap supports the full grammar 'butt | round | square'"/>
  </metadata>
  <g id="target"></g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/parsing-testcommon.js"/>
  <script><![CDATA[

test_valid_value("stroke-linecap", "butt");
test_valid_value("stroke-linecap", "round");
test_valid_value("stroke-linecap", "square");

  ]]></script>
</svg>
